Fix: renamed no-magic-number to no-magic-numbers (fixes #4053)
