+++ lib/rules/no-magic-numbers.js
